requirements.txt
numpy==1.24.3
pygame==2.5.2
tensorflow==2.14.0
matplotlib==3.7.3
pandas==2.0.3
scikit-learn==1.3.1
gymnasium==0.29.1
opencv-python==4.8.1.78
tqdm==4.66.1
flask==2.3.3
pillow==10.0.0 